scaffolding(portafolio_sofi_may2026): split magnet files + template hooks + S18 fixture
S1: partition apps/lineas/{models,urls}, apps/cuadrillas/{models,urls}, apps/campo/urls
S2: template include hooks (ignore missing) for filtro_semestre + filtro_estado
S3: fixtures/Programacion_S18_real.xlsx (download from #48 attachment)

Refs: #100 #101 #102 #104 #105 #48 #72

Nota F2: blueprint nombraba templates linea_detalle.html y cuadrilla_list.html;
los archivos reales son detalle.html y lista.html — se modificaron los reales.

🤖 Generado por F2 del skill /modulo
