+++ templates/campo/avance_registrar.html
@@ -11,6 +11,9 @@
         <p class="text-gray-600 dark:text-gray-400">Registra el estado de ejecución de los vanos de tu línea asignada</p>
     </div>
 
+    {# B2.1 hook (Sofi mayo 2026) — segmentar grid por semestre S1/S2/TA. Sub-feature crea templates/lineas/_filtro_semestre.html. #}
+    {% include 'lineas/_filtro_semestre.html' ignore missing %}
+
     <!-- Error Message -->
     {% if error %}
     <div class="bg-red-50 dark:bg-red-900/30 border border-red-200 dark:border-red-800 rounded-lg p-4">
