+++ templates/lineas/detalle.html
@@ -14,6 +14,8 @@
 
 {% block content %}
 <div class="space-y-6">
+    {# B2.1 hook (Sofi mayo 2026) — sub-feature crea templates/lineas/_filtro_semestre.html. Mientras no exista, Django no falla por `ignore missing`. #}
+    {% include 'lineas/_filtro_semestre.html' ignore missing %}
     <!-- Header -->
     <div class="flex flex-col sm:flex-row justify-between items-start sm:items-center gap-4">
         <div class="flex items-center gap-4">
